+++ tests/test_surround.lua
@@ -120,6 +120,7 @@ T['setup()']['creates `config` field'] = function()
   expect_config('mappings.update_n_lines', 'sn')
   expect_config('mappings.suffix_last', 'l')
   expect_config('mappings.suffix_next', 'n')
+  expect_config('respect_selection_type', false)
   expect_config('search_method', 'cover')
 end
 
@@ -150,6 +151,7 @@ T['setup()']['validates `config` argument'] = function()
   expect_config_error({ mappings = { suffix_last = 1 } }, 'mappings.suffix_last', 'string')
   expect_config_error({ mappings = { suffix_next = 1 } }, 'mappings.suffix_next', 'string')
   expect_config_error({ n_lines = 'a' }, 'n_lines', 'number')
+  expect_config_error({ respect_selection_type = 1 }, 'respect_selection_type', 'boolean')
   expect_config_error({ search_method = 1 }, 'search_method', 'one of')
 end
 
@@ -391,6 +393,79 @@ T['Add surrounding']['works in line and block Visual mode'] = function()
   validate_edit({ 'aaa', 'bbb' }, { 1, 0 }, { '(aaa', 'bbb)' }, { 1, 1 }, type_keys, '<C-v>j$', 'sa', ')')
 end
 
+--stylua: ignore
+T['Add surrounding']['respects `config.respect_selection_type` in linewise mode'] = function()
+  child.lua('MiniSurround.config.respect_selection_type = true')
+
+  local validate = function(before_lines, before_cursor, after_lines, after_cursor, selection_keys)
+    validate_edit(before_lines, before_cursor, after_lines, after_cursor, type_keys, selection_keys, 'sa', ')')
+  end
+
+  -- General test in Visual mode
+  validate({ 'aaa' }, { 1, 0 }, { '(', '\taaa', ')' }, { 2, 1 }, 'V')
+
+  -- Correctly computes indentation
+  validate({ 'aaa',   ' bbb', '  ccc' }, { 2, 0 }, { 'aaa', ' (',      '\t bbb', '\t  ccc', ' )' },  { 3, 2 }, 'Vj')
+  validate({ ' aaa',  '',     ' bbb' },  { 1, 0 }, { ' (',  '\t aaa',  '',       '\t bbb',  ' )' },  { 2, 2 }, 'V2j')
+  validate({ '  aaa', ' ',    '  bbb' }, { 1, 0 }, { '  (', '\t  aaa', '\t ',    '\t  bbb', '  )' }, { 2, 3 }, 'V2j')
+
+  -- Handles empty/blank lines
+  validate({ '  aaa', '', ' ', '  bbb' }, { 1, 0 }, { '  (', '\t  aaa', '', '\t ', '\t  bbb', '  )' }, { 2, 3 }, 'V3j')
+
+  validate({ '',  '  aaa', '' },  { 1, 0 }, { '  (', '',    '\t  aaa', '',    '  )' }, { 2, 0 }, 'V2j')
+  validate({ ' ', '  aaa', ' ' }, { 1, 0 }, { '  (', '\t ', '\t  aaa', '\t ', '  )' }, { 2, 1 }, 'V2j')
+
+  -- Works with different surroundings
+  validate_edit({ 'aaa' }, { 1, 0 }, { 'ff(', '\taaa', ')' }, { 2, 1 }, type_keys, 'V', 'sa', 'f', 'ff<CR>')
+
+  -- General test in Operator-pending mode
+  validate_edit({ 'aaa' }, { 1, 0 }, { '(', '\taaa', ')' }, { 2, 1 }, type_keys, 'sa', 'ip', ')')
+
+  -- Respects `expandtab`
+  child.o.expandtab = true
+  child.o.shiftwidth = 3
+  validate({ 'aaa' }, { 1, 0 }, { '(', '   aaa', ')' }, { 2, 3 }, 'V')
+end
+
+--stylua: ignore
+T['Add surrounding']['respects `config.respect_selection_type` in blockwise mode'] = function()
+  -- NOTE: this doesn't work with mix of multibyte and normal characters,
+  -- as well as outside of text lines.
+  child.lua('MiniSurround.config.respect_selection_type = true')
+
+  local validate = function(before_lines, before_cursor, after_lines, after_cursor, selection_keys)
+    validate_edit(before_lines, before_cursor, after_lines, after_cursor, type_keys, selection_keys, 'sa', ')')
+  end
+
+  -- General test in Visual mode
+  validate({ 'aaa', 'bbb' }, { 1, 1 }, { 'a(a)a', 'b(b)b' }, { 1, 2 }, '<C-v>j')
+  validate({ 'aaaa', 'bbbb' }, { 1, 1 }, { 'a(aa)a', 'b(bb)b' }, { 1, 2 }, '<C-v>jl')
+
+  -- Works on single line
+  validate({ 'aaaa' }, { 1, 1 }, { 'a(aa)a' }, { 1, 2 }, '<C-v>l')
+
+  -- Works when selection is created in different directions
+  validate({ 'aaaa', 'bbbb' }, { 1, 2 }, { 'a(aa)a', 'b(bb)b' }, { 1, 2 }, '<C-v>jh')
+  validate({ 'aaaa', 'bbbb' }, { 2, 1 }, { 'a(aa)a', 'b(bb)b' }, { 1, 2 }, '<C-v>kl')
+  validate({ 'aaaa', 'bbbb' }, { 2, 2 }, { 'a(aa)a', 'b(bb)b' }, { 1, 2 }, '<C-v>kh')
+
+  -- Works with different surroundings
+  validate_edit({ 'aaa', 'bbb' }, { 1, 1 }, { 'aff(a)a', 'bff(b)b' }, { 1, 4 }, type_keys, '<C-v>j', 'sa', 'f', 'ff<CR>')
+
+  -- General test in Operator-pending mode
+  set_lines({ 'aaaaa', 'bbbbb' })
+
+  -- - Create mark to be able to perform non-trival movement
+  set_cursor(2, 3)
+  type_keys('ma')
+
+  set_cursor(1, 1)
+  type_keys('sa', '<C-v>', '`a', ')')
+  -- - As motion is end-exclusive, it registers end mark one column short.
+  eq(get_lines(), { 'a(aa)aa', 'b(bb)bb' })
+  eq(get_cursor(), { 1, 2 })
+end
+
 T['Add surrounding']['validates single character user input'] = function()
   validate_edit({ ' aaa ' }, { 1, 1 }, { ' aaa ' }, { 1, 1 }, type_keys, 'sa', 'iw', '<C-v>')
   eq(get_latest_message(), '(mini.surround) Input must be single character: alphanumeric, punctuation, or space.')
@@ -580,6 +655,40 @@ T['Delete surrounding']['works with dot-repeat'] = function()
   set_cursor(1, 5)
   type_keys('.')
   eq(get_lines(), { 'aaa bbb' })
+end
+
+--stylua: ignore
+T['Delete surrounding']['respects `config.respect_selection_type` in linewise mode'] = function()
+  child.lua('MiniSurround.config.respect_selection_type = true')
+
+  local validate = function(before_lines, before_cursor, after_lines, after_cursor)
+    validate_edit(before_lines, before_cursor, after_lines, after_cursor, type_keys, 'sd', ')')
+  end
+
+  -- General test
+  validate({ '(', '\taaa', ')' }, { 2, 0 }, { 'aaa' }, { 1, 0 })
+
+  -- Works when cursor is on any part of region
+  validate({ '(', '\taaa', ')' }, { 1, 0 }, { 'aaa' }, { 1, 0 })
+  validate({ '(', '\taaa', ')' }, { 3, 0 }, { 'aaa' }, { 1, 0 })
+
+  -- Correctly applies when it should
+  validate({ '(',   '\t\taaa', '\tbbb', ')' },   { 2, 2 }, { '\taaa', 'bbb' }, { 1, 1 })
+  validate({ '  (', '\t\taaa', '\tbbb', ')  ' }, { 2, 2 }, { '\taaa', 'bbb' }, { 1, 1 })
+
+  -- Correctly doesn't apply when it shouldn't
+  validate({ 'aaa',  '  ()  ', 'bbb' },  { 2, 2 }, { 'aaa', '    ',  'bbb' }, { 2, 2 })
+  validate({ 'aaa(', '\tbbb',  ')' },    { 2, 2 }, { 'aaa', '\tbbb', '' },    { 1, 2 })
+  validate({ '(',    '\tbbb',  ')ccc' }, { 2, 2 }, { '',    '\tbbb', 'ccc' }, { 1, 0 })
+
+  -- Correctly dedents
+  validate({ '(', 'aaa', ')' }, { 2, 0 }, { 'aaa' }, { 1, 0 })
+
+  child.o.shiftwidth = 3
+  validate({ '(', '    aaa', ')' }, { 2, 0 }, { ' aaa' }, { 1, 1 })
+
+  child.o.expandtab = true
+  validate({ '(', '    aaa', ')' }, { 2, 0 }, { ' aaa' }, { 1, 1 })
 end
 
 T['Delete surrounding']['works in extended mappings'] = function()